Move domReady module to its own package.
