+++ packages/domReady/src/index.js
